More comments for sample generator, more generalized soft-coding for easier user tweaking

+++ Extras/Examples/sampleData/generateSampleData.py
@@ -1,7 +1,49 @@
 '''
-Generate/overwrite sample data to test oscaar with.
+Generate a simulated transit observation to test oscaar with.
+Three stars in the image from left to right are the target, 
+comparison star A, and comparison star B. 
 
-Requires that the C library is compiled, see instructions: 
+WARNING: If you want to keep the contents of the 
+`OSCAAR/Extras/Examples/sampleData/images` directory, save a 
+copy of it with a different name. This script will overwrite 
+any directory named `images` within this directory.
+
+From the oscaar Wiki:
+    (https://github.com/OSCAAR/OSCAAR/wiki/Generating-Sample-Data)
+    A transit of the target star will occur in the middle of the 
+    simulated observations, and we use these data to see how our 
+    differential photometry routines perform. Though the simulated 
+    "stars" are exceedingly square approximations of what real 
+    stellar image shapes look like in real observations, some 
+    important sources of signal and noise are included in the 
+    simulation in order to test oscaar in as close to battle-field 
+    conditions as possible. We simulate dark current, sky background, 
+    stars of different magnitudes and appropriate photon noise for 
+    each of those components. We also jitter the centroids of the 
+    stars around randomly in small two dimensional shifts to 
+    demonstrate the capabilities of oscaar's stellar centroid 
+    tracking algorithm. Below is one such simulated image with the 
+    star magnitudes significantly decreased so as to portray the 
+    shot noise in the background as well as in the stars in the 
+    same colormapping.
+
+    ...
+
+    Here are some `init.par` and `observatory.par` parameters that 
+    should be set to particular values to achieve successful 
+    differential photometry from the simulated data sets. From 
+    `init.par`:
+        Smoothing Constant = 3
+        Tracking Zoom = 15
+        Ingress = 2013-05-15;10:06:30
+        Egress = 2013-05-15;11:02:35 
+    And from `observatory.par`,
+        Exposure Time Keyword: JD
+
+Requires that the C library for generating transit light curves
+is compiled. If it's not, change directories to `sampleData/c`
+and execute `python setup.py build_ext --inplace`. 
+See the detailed instructions on the oscaar wiki: 
 https://github.com/OSCAAR/OSCAAR/wiki/Generating-Sample-Data
 
 Core developer: Brett Morris
@@ -9,26 +51,39 @@ Core developer: Brett Morris
 
 import numpy as np
 import pyfits
-from matplotlib import pyplot as plt
-import generateModelLC as generateModel
 from shutil import rmtree
 from os import mkdir
 from glob import glob
+from matplotlib import pyplot as plt
 
-plotModel = False
-NdataImages = 200          ## Number of data images to generate
+## This import assumes that the "generateModelLC.py"
+##   script is available in the same directory as 
+##   generateSampleData.py. 
+import generateModelLC as generateModel 
+
+
+#################################################################
+## Tweak these parameters, if you like!
+NdataImages = 200        ## Number of data images to generate
 NdarkImages = 3          ## Number of dark frames to generate
 NflatImages = 3          ## Number of flat fields to generate
-imageDimensionX = 120     ## pixel dimensions of each image
+flatFieldCounts = 20000  ## Approx counts per pixel in flat field
+imageDimensionX = 120    ## Pixel dimensions of each image
 imageDimensionY = 40
-starDimensions = 4       ## pixel dimensions of the stars
-skyBackground = 500      ## background counts from sky brightness
-darkBackground = 100     ## background counts from detector
+starDimensions = 4       ## Pixel dimensions of the stars
+skyBackground = 500      ## Background counts from sky brightness
+darkBackground = 100     ## Background counts from detector
+targetFluxOOT = 10000    ## Flux (in counts) from each pixel of the unocculted target star (out-of-transit)
+relativeFluxCompA = 0.85 ## Flux from comp A relative to target
+relativeFluxCompB = 0.95 ## Flux from comp B relative to target
+plotModel = False        ## Plot the injected transit light curve
+createMasterFlatNow = True  ## Use oscaar to create a master flat from the freshly generated flat frames
+## Isn't it nice to control the signal to noise?
+#################################################################
 
 ## Delete `images` directory, if there is one, and
 ##      make a fresh one.
-if len(glob('images')) > 0:
-	rmtree('images')
+if len(glob('images')) > 0: rmtree('images')
 mkdir('images')
 
 ## Pixel positions of the stars (x,y)
@@ -37,13 +92,12 @@ compAX = [60-starDimensions/2,60+starDimensions/2]
 compBX = [100-starDimensions/2,100+starDimensions/2]
 starsY = [imageDimensionY/2-starDimensions/2,imageDimensionY/2+starDimensions/2]
 
-## Set times, model params
-## Ingress: 2013-05-15;10:06:30; egress: 2013-05-15;11:02:35
-jd0 = 2456427.88890
+## Set ingress/egress times, transiting system parameters 
+## In GD: Ingress = 2013-05-15;10:06:30; egress = 2013-05-15;11:02:35
+jd0 = 2456427.88890 
 exposureTime = 45/(60*60*24.) ## Convert s -> hr
-
-# [p,ap,P,i,gamma1,gamma2,e,longPericenter,t0,percentOfOrbit]
 times = np.arange(jd0,jd0+exposureTime*NdataImages,exposureTime)
+# [p,ap,P,i,gamma1,gamma2,e,longPericenter,t0]
 modelParams = [ 0.1179, 14.71, 1.580400, 90.0, 0.23, \
                 0.30, 0.00, 0.0, np.mean(times,dtype=np.float64)]
 modelLightCurve = generateModel.simulateLC(times,modelParams)
@@ -52,45 +106,57 @@ if plotModel:
 	ax1 = fig.add_subplot(111)
 	def format_coord(x, y):
 		'''Function to also give data value on mouse over with imshow.'''
-		col = int(x+0.5)
-		row = int(y+0.5)
-		return 'JD=%1.8f, Relative Flux=%1.8f' % (x, y)
-	plt.plot(times,modelLightCurve)
+		return 'JD=%1.6f, Flux=%1.6f' % (x, y)
+	ax1.set_xlabel('Time (JD)')
+	ax1.set_ylabel('Relative Flux')
+	ax1.set_title('Injected Light Curve')
+	ax1.plot(times,modelLightCurve)
 	ax1.format_coord = format_coord
 	plt.show()
 
+## For producing random arrays, initialize reference arrays with the proper shapes
+imageShapedMatrix = np.zeros([imageDimensionY,imageDimensionX])
+starShapedMatrix = np.zeros([starDimensions,starDimensions])
+
 ## Simulate dark frames with shot noise
 for i in range(NdarkImages):
-    darkFrame = darkBackground + np.random.normal(np.zeros([imageDimensionY,imageDimensionX]),np.sqrt(darkBackground))
+    darkFrame = darkBackground + np.random.normal(imageShapedMatrix,np.sqrt(darkBackground))
     darkFrame = np.require(darkFrame,dtype=int)   ## Require integer counts
     pyfits.writeto('images/simulatedImg-'+str(i).zfill(3)+'d.fits',darkFrame)
 
 ## Simulate ideal flat frames (perfectly flat)
 for i in range(NflatImages):
     ## Flats will be completely flat -- ie, we're pretending that we have a 
-    ##      perfect optical path with no variations.
-    flatField = np.ones([imageDimensionY,imageDimensionX],dtype=np.int) ## Require integer counts
+    ##      perfect optical path with no spatial flux variations.
+    flatField = np.zeros([imageDimensionY,imageDimensionX]) +  flatFieldCounts
+    flatField = np.require(flatField,dtype=int)## Require integer counts
     pyfits.writeto('images/simulatedImg-'+str(i).zfill(3)+'f.fits',flatField)
 
+
+## Create master flat now using oscaar's standard flat maker
+if createMasterFlatNow:
+    import oscaar
+    flatPaths = glob('images/simulatedImg-???f.fits')
+    flatDarkPaths = glob('images/simulatedImg-???d.fits')   ## Use the same darks
+    masterFlatSavePath = 'images/masterFlat.fits'   ## Where to save the master
+    oscaar.standardFlatMaker(flatPaths,flatDarkPaths,masterFlatSavePath,plots=False)
+
+
 ## Create data images
-for i in range(0,NdataImages):
+for i in range(NdataImages):
     ## Produce image with sky and dark background with simulated photon noise for each source
-    simulatedImage = darkBackground +\
-        np.random.normal(np.zeros([imageDimensionY,imageDimensionX]),np.sqrt(darkBackground)) + skyBackground +\
-        np.random.normal(np.zeros([imageDimensionY,imageDimensionX]),np.sqrt(skyBackground))
+    simulatedImage = darkBackground + skyBackground +\
+        np.random.normal(imageShapedMatrix,np.sqrt(darkBackground)) +\
+        np.random.normal(imageShapedMatrix,np.sqrt(skyBackground))
     
     ## Create two box-shaped stars with simulated photon noise
-    targetBrightness = 3*3000*modelLightCurve[i]  ## Scale brightness with the light curve
-    target = targetBrightness +\
-        np.random.normal(np.zeros([starDimensions,starDimensions]),np.sqrt(targetBrightness))
+    targetBrightness = targetFluxOOT*modelLightCurve[i]  ## Scale brightness with the light curve
+    target = targetBrightness + np.random.normal(starShapedMatrix,np.sqrt(targetBrightness))
     
-    compBrightnessA = 3*2500
-    compBrightnessB = 3*2700
-    compA = compBrightnessA +\
-        np.random.normal(np.zeros([starDimensions,starDimensions]),np.sqrt(compBrightnessA))
-    compB = compBrightnessB +\
-        np.random.normal(np.zeros([starDimensions,starDimensions]),np.sqrt(compBrightnessB))
-    
+    compBrightnessA = targetFluxOOT*relativeFluxCompA
+    compBrightnessB = targetFluxOOT*relativeFluxCompB
+    compA = compBrightnessA + np.random.normal(starShapedMatrix,np.sqrt(compBrightnessA))
+    compB = compBrightnessB + np.random.normal(starShapedMatrix,np.sqrt(compBrightnessB))
     
     ## Add stars onto the simulated image with some position jitter
     randomPositionJitterX = np.sign(np.random.uniform(-1,1))	## +/- 1 pixel stellar centroid position jitter
@@ -100,7 +166,6 @@ for i in range(0,NdataImages):
     simulatedImage[starsY[0]+randomPositionJitterY:starsY[1]+randomPositionJitterY,compBX[0]+randomPositionJitterX:compBX[1]+randomPositionJitterX] += compB
 
     ## Force counts to integers, save.
-    #simulatedImage = np.transpose(simulatedImage)
     simulatedImage = np.require(simulatedImage,dtype=int)   ## Require integer counts before save
 
     header = pyfits.Header()
